[codex:orchestration] extract current export/acceptance projection slice
